Fetch ai message from server
