+++ frontend/src/features/chat/message-component.tsx
